+++ crates/criticalup-core/src/download_server_client.rs
@@ -1,6 +1,8 @@
 // SPDX-FileCopyrightText: The Ferrocene Developers
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+use std::time::Duration;
+
 use crate::config::Config;
 use crate::envvars;
 use crate::errors::{DownloadServerError, Error};
@@ -31,6 +33,13 @@ impl DownloadServerClient {
         let retry_policy = ExponentialBackoff::builder().build_with_max_retries(CLIENT_MAX_RETRIES);
         let client = reqwest::ClientBuilder::new()
             .user_agent(config.whitelabel.http_user_agent)
+            .read_timeout(Duration::from_secs(90))
+            .connect_timeout(Duration::from_secs(90))
+            .pool_idle_timeout(Duration::from_secs(90))
+            // In rare cases we were encountering a hang in networking in Docker-in-Docker
+            // `docker buildx build` situations. This workaround seems to help.
+            // ref: https://github.com/hyperium/hyper/issues/2312#issuecomment-778005053
+            .pool_max_idle_per_host(0)
             .build()
             .expect("failed to configure http client");
         let client = ClientBuilder::new(client)